polya

+++ src/test/test4.hpp
@@ -1130,7 +1130,7 @@ namespace test4_namespace
 		//10001000
 		char x=129;
 		const int N=8;
-		std::cout<<rotateShiftLeft(x,N)<<std::endl;
+		std::cout<<(int)rotateShiftLeft(x,N)<<std::endl;
 	}
 	void test_out()
 	{
